edit stuff in dataset separated
